Move bootstrap to the app
