+++ src/bootstrap.js
@@ -1,11 +1,11 @@
 import angular from 'angular';
-import module from './module';
+import module from './app/module';
 
 export default function($document) {
   'ngInject';
 
   angular.element($document).ready(() => {
-    angular.bootstrap($document, [module.name], {
+    angular.bootstrap($document.find('body'), [module.name], {
       strictDi: true
     });
   });
